Add IPv6 Jumbogram support (RFC 2675)
Reviewed-on: https://code.wireshark.org/review/9968
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>

+++ epan/dissectors/packet-udp.h
@@ -36,7 +36,7 @@ typedef struct _e_udphdr {
   guint16 uh_sport;
   guint16 uh_dport;
   guint16 uh_ulen;
-  guint16 uh_sum_cov;
+  guint32 uh_sum_cov;
   guint16 uh_sum;
   guint32 uh_stream; /* this stream index field is included to help differentiate when address/port pairs are reused */
   address ip_src;
